Tests: improve setup and cleanup
Tests often need to create temporary files and directories. These are
usually cleaned at the end but if the test fails, the tests doesn't
happen. Also, since Templaar is directory-based, it would be better to
execute each test in a separate directory.

To improve the test experience, add a new 'Test' struct which:
- sets up the EDITOR env var
- creates a test-specific directory
- pre-initializes necessary files (with given contents) and directories
  inside the test directory
- changes to the test directory
- implements the std::ops::Drop trait to change back to the original
  dir and remove the test dir

Creating a 'Test' object inside the test simplifies the initialization
and makes sure that everything is properly cleaned, even if the test
fails (thanks to the Drop trait).