Add --global option to the 'new' command
When the option is used, the template will be stored in a global
directory ~/.config/templaar. This is useful for templates that can be
used from anywhere. Support for 'take' command will follow.

Note that the template files in the global directory are not hidden as
it does not make sense (directory would seem empty). On the contrary, it
is more useful to see all available templates by just listing the
directory.