Introduce directory templates from multiple files
Allow multiple files to be specified `templaar new -f`. When this
happens, the template is created as a directory with all the specified
files under their original names (the name of the directory matches the
template name).

For example:

    $ templaar new templ -f a.txt b.txt

creates a new directory called `.templ.aar` containing files a.txt and
b.txt with their original content.

Finally, the directory is opened with the default editor to allow
further modification of the files.

Note: it is now necessary to specify the template name argument *before*
the -f argument, otherwise the name would be parsed as one of the files.